--- CoinNames.js
@@ -1,5 +1,0 @@
-/**
- * CoinNames Currently Displayed on GasFee Table
- */
-
- module.exports = ["Eth" , "Ftm", "Matic", "Bnb", "Icp", "Dot", "Sol", "Car"];
