+++ Data/Data.js
@@ -1,0 +1,4 @@
+/**
+ * CoinNames Currently Displayed on GasFee Table
+ */
+module.exports = ["Eth" , "Ftm", "Matic", "Bnb", "Icp", "Dot", "Sol", "Car"];
