+++ index.js
@@ -1,66 +1,29 @@
 const express = require("express");
 const { createServer } = require("http");
 const { Server } = require("socket.io");
-const axios = require('axios');
-require('dotenv').config();
-const bot = require("./bot");
-
-/**
- * Emits an event containing the current data of a particular coin
- * @param {string} coinName  - Name of the coin to send data of 
- * @param {socket} socket - Socket.io socket
- */
-function sendCoinUpdate(coinName, socket){
-    const data = bot.getCoinData(coinName);
-    const event = coinName + "Update";
-    socket.emit(event, data);
-}
-
-/**
- * Subscirbes the socket to listen to GasFee Table updates
- * Sets intervals that periodically send out events with respect to certain coins
- * @param {socket} socket 
- */
-function subscribeSocketGas(socket){
-    const CoinList = bot.getCoinList();
-    for(let i = 0 ; i < CoinList.length; i ++){
-        let currCoin = CoinList[i];
-        setInterval(()=>{
-            sendCoinUpdate(currCoin, socket);
-        }, 5000)
-    }
-}
+const bot = require("./Data/databot");
+const handleSocket = require("./Client/ws"); //websocket api
+const gasRoute = require("./Client/restapi");
+bot.config();
+bot.start();
 
 
-
-
-
-
-
-
-
-
-//Server/Socket.io initialization 
+//Initializing https server and routes
 const app = express();
+app.use("/gas", gasRoute);
 const httpServer = createServer(app);
 const io = new Server(httpServer, {
   cors: {
     origin: true,
   },
 });
-bot.config();
-if(process.argv[2] == "Test")
-{
- 
-}
 
-//Socket.io event updates
+//Initializing socket.io events
 io.on("connection", (socket) => {
-    bot.start();
-    subscribeSocketGas(socket);
+    handleSocket(socket);
 })
 
-
+//Listening on port 5000
 httpServer.listen(5000);
